requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
ta>=0.10.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-telegram-bot>=20.7
schedule>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0 